pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TaskManagement"
version = "0.1.0"
requires-python = "== 3.11.3"
dependencies = [
    "fastapi==0.115.12",
    "Flask==3.1.0",
    "google-genai==1.9.0",
    "python-dotenv==1.1.0",
    "uvicorn==0.34.0",
    "pydantic==2.11.1"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
]

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
